requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
cls==2019.4.13
Django==3.0.5
django-crispy-forms==1.9.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
public==2019.4.13
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.42
requests==2.23.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.0.1
xhtml2pdf==0.2.4
